A lot of changes, disabled the visitor pattern (ready to remove it completly) as it isn't used atm; disabled goToPos() method of BasicElement because this will be replaced with childElementAt(). This implementation will be simpler and independend from the FormulaCursor class which will be reworked too, in order to put all available cursor functionality in just one class, instead of spreading it over all element classes. Made ContextStyle class inheriting from KoZoomHandler instead of KoTextZoomHandler, copied the relevant methods from KoTextZoomHandler, it is only temporary as long as the new rendering class isn't finished. But this removed a lot of the warnings that came up with the transition from int to float values. Imporved Api dox here and there, renamed some things, other simple and minor changes... Every thing compiles fine and links too but don't expect it to actually work.
svn path=/trunk/koffice/; revision=552162

+++ kformula/kformulawidget.h
@@ -82,7 +82,7 @@ protected slots:
     /**
      * The formula has changed and needs to be redrawn.
      */
-    void slotFormulaChanged(int width, int height);
+    void slotFormulaChanged( double width, double height );
 
     void slotCursorChanged(bool visible, bool selecting);
